finish dpeloy powershell script

+++ src/deploy.ps1
@@ -1,10 +1,12 @@
-﻿
-Remove-Item .\deploy -Recurse -Confirm:$false
-Remove-Item .\www -Recurse -Confirm:$false
-Remove-Item .\deploy.zip -Confirm:$false
+﻿Remove-Item .\deploy -Recurse -Confirm:$false
+if (Test-Path .\deploy.zip) {
+    Remove-Item .\deploy.zip -Confirm:$false
+}
 
 npm run build
 
+Get-ChildItem .\deploy\www\browser\* -Recurse | Move-Item -Destination .\deploy\www\
+Copy-Item .\backend-player\README.md .\deploy\README.md -Confirm:$false
 Compress-Archive -Path .\deploy\* -DestinationPath .\deploy.zip
 
 Write-Host "----------------------------------"
@@ -14,9 +16,10 @@ $confirmation = Read-Host "Deploy to bin-folder y(es) or n(o)"
 if ($confirmation -eq 'y') {
     Remove-Item ..\..\bin\nodejs\deploy.zip -Confirm:$false
     Copy-Item .\deploy.zip ..\bin\nodejs\deploy.zip -Confirm:$false
-    Write-Host "Finished with copy"
-    }
-else
-    {
+    Remove-Item .\deploy.zip -Confirm:$false
+    Write-Host "Finished with copy, removed local zip."
+}
+else {
     Write-Host "Finished without copy"
-    }
+}
+Remove-Item .\deploy -Recurse -Confirm:$false
